refactor(agents): rename data-collection agents to extractors
Extractor files make zero Claude API calls — calling them "agents"
implies LLM-driven autonomy they do not have. Recommendation Agent
and Orchestrator keep the _agent suffix as they are the only two
components that call Claude.

Renames:
- frontend_sentry_agent.py  → frontend_sentry_extractor.py
- backend_sentry_agent.py   → backend_sentry_extractor.py (pending, doc only)
- render_logs_agent.py      → render_logs_extractor.py (pending, doc only)
- github_agent.py           → github_extractor.py (pending, doc only)
- codebase_agent.py         → codebase_extractor.py (pending, doc only)

Updates all references in agent-architecture.md, agent-execution-plan.md,
mcp-vs-agents.md, dt13_agent_observability.md, README.md, and test file.
