+++ agents/frontend_sentry_extractor.py
